+++ python_search/ps_llm/t5/next_item_prompt.py
@@ -2,7 +2,8 @@ class PromptBuilder:
     """
     Builds the prompt for train or inference
     """
-    PROMPT_START = "predict the next key given this history: "
+    PROMPT_START = "Predict the next key given this history: "
+
     def build_prompt_inference(self, history):
         prompt = f"{self.PROMPT_START} "