+++ python_search/ps_llm/tasks/next_item_predictor.py
@@ -9,6 +9,7 @@ class NextItemPredictor(BaseDataTask):
 
     def prompt(self, history):
         return PromptBuilder().build_prompt_inference(history)
+
     def build_dataset(self, skip_clean=False):
         """
         Generates the dataset and writes it to disk