+++ python_search/share_entry.py
@@ -15,7 +15,7 @@ class ShareEntry:
             raise Exception(f"Entry {key} not found")
 
         entry = Entry(key, self._entries[key])
-        result = f"{entry.key}: {entry.get_only_content()}"
+        result = f"{entry.key}: {entry.get_content_str()}"
 
         Clipboard().set_content(result, enable_notifications=True, notify=True)